refactor: Move buildApp and buildStudio logic to cli-build package
