+++ packages/@sanity/cli/src/actions/dev/startStudioDevServer.ts
@@ -1,6 +1,10 @@
 import {styleText} from 'node:util'
 
-import {checkStudioDependencyVersions} from '@sanity/cli-build/_internal/build'
+import {
+  checkRequiredDependencies,
+  checkStudioDependencyVersions,
+  compareDependencyVersions,
+} from '@sanity/cli-build/_internal/build'
 import {getLocalPackageVersion, isInteractive} from '@sanity/cli-core'
 import {confirm, logSymbols, spinner} from '@sanity/cli-core/ux'
 import {parse as semverParse} from 'semver'
@@ -9,10 +13,8 @@ import {startDevServer} from '../../server/devServer.js'
 import {gracefulServerDeath} from '../../server/gracefulServerDeath.js'
 import {getProjectById} from '../../services/projects.js'
 import {getAppId} from '../../util/appId.js'
-import {compareDependencyVersions} from '../../util/compareDependencyVersions.js'
 import {getPackageManagerChoice} from '../../util/packageManager/packageManagerChoice.js'
 import {upgradePackages} from '../../util/packageManager/upgradePackages.js'
-import {checkRequiredDependencies} from '../build/checkRequiredDependencies.js'
 import {shouldAutoUpdate} from '../build/shouldAutoUpdate.js'
 import {devDebug} from './devDebug.js'
 import {getDashboardAppURL} from './getDashboardAppUrl.js'
